+++ src/pages/GamePage/components/ShipsMenu/ShipsMenu.tsx
@@ -1,46 +1,19 @@
-// todo: разобраться с импортом
-//  Ошибка: Unable to resolve path to module 'images/ships/ship1.png'
-/* eslint-disable */
-import ship1Img from 'images/ships/ship1.png';
-import ship2Img from 'images/ships/ship2.png';
-import ship3Img from 'images/ships/ship3.png';
-import ship4Img from 'images/ships/ship4.png';
-/* eslint-enable */
+import { useMemo } from 'react';
+import { getCurrentShips } from './helpers/getCurrentShips';
+import { CONTAINER_WIDTH, CONTAINER_HEIGHT } from './data';
 import styles from './ShipsMenu.scss';
-import {
-    CONTAINER_WIDTH,
-    CONTAINER_HEIGHT,
-    IMG_WIDTH,
-    MARGIN,
-    SHIP_1_HEIGHT,
-} from './data';
 import type { Props } from './types';
 
-export const ShipsMenu = ({ ships }: Props) => (
-    <div
-        className={styles.container}
-        style={{ width: CONTAINER_WIDTH, height: CONTAINER_HEIGHT }}
-    >
-        {ships.map(({ id, type, visible }, index) => {
-            if (visible) return '';
+export const ShipsMenu = ({ ships }: Props) => {
+    const currentShips = useMemo(() => getCurrentShips(ships), [ships]);
 
-            let src = '';
-            if (type === 1) src = ship1Img;
-            if (type === 2) src = ship2Img;
-            if (type === 3) src = ship3Img;
-            if (type === 4) src = ship4Img;
-
-            let bottom = 0;
-            let left = (IMG_WIDTH + MARGIN) * index;
-
-            // рендер однопалубных кораблей друг над другом
-            if (index > 7) {
-                bottom = MARGIN + SHIP_1_HEIGHT;
-                left = (IMG_WIDTH + MARGIN) * (index - 2);
-            }
-
-            return (
-                <img
+    return (
+        <div
+            className={styles.container}
+            style={{ width: CONTAINER_WIDTH, height: CONTAINER_HEIGHT }}
+        >
+            {currentShips.map(({ id, src, bottom, left }) => (
+    <img
                     alt=""
                     key={id}
                     data-test-id={id}
@@ -48,7 +21,7 @@ export const ShipsMenu = ({ ships }: Props) => (
                     src={src}
                     style={{ bottom, left }}
                 />
-            );
-        })}
-    </div>
-);
+            ))}
+        </div>
+    );
+};
